Handle log subscriber symbol logger level
24189390074049ca16da003dc796f1f64622173a introduced a change to
ActiveSupport::LogSubscriber expecting its Logger’s level to
always be an integer. It would be helpful for using custom loggers
if we could continue to use symbol based levels.

This PR memoizes the logger level on the LogSubscriber class in a
log_level function. The memoize step retrieves the relevant
integer constant if the logger level is a symbol.