Handle log subscriber symbol logger level
2418939 introduced a change to ActiveSupport::LogSubscriber
expecting its logger’s level to always be an integer. It
would be helpful when using custom loggers to be able to
continue using symbol based levels.

This PR memoizes the logger level on the LogSubscriber class in a
log_level function. The memoize step retrieves the relevant
integer constant if the logger level is a symbol.